move content to group folder
