+++ .hlxignore
@@ -5,3 +5,4 @@ LICENSE
 package.json
 package-lock.json
 test/*
+tools/importer/
