+++ 404.html
@@ -9,7 +9,7 @@
   </script>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta property="og:title" content="Page not found">
-  <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
+  <script src="/group/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
     window.addEventListener('load', () => {
       if (document.referrer) {
@@ -27,10 +27,10 @@
     });
   </script>
   <script type="module">
-    import { sampleRUM } from '/scripts/aem.js';
+    import { sampleRUM } from '/group/scripts/aem.js';
     sampleRUM('404', { source: document.referrer });
   </script>
-  <link rel="stylesheet" href="/styles/styles.css">
+  <link rel="stylesheet" href="/group/styles/styles.css">
   <style>
     main.error {
       min-height: calc(100vh - var(--nav-height));
@@ -46,7 +46,7 @@
       font-family: monospace;
     }
   </style>
-  <link rel="stylesheet" href="/styles/lazy-styles.css">
+  <link rel="stylesheet" href="/group/styles/lazy-styles.css">
 </head>
 
 <body>
@@ -58,7 +58,7 @@
       </svg>
       <h2 class="error-message">Page Not Found</h2>
       <p class="button-container">
-        <a href="/" class="button secondary error-button-home">Go home</a>
+        <a href="/group" class="button secondary error-button-home">Go home</a>
       </p>
     </div>
   </main>
