--- blocks/cards/cards.js
@@ -1,46 +1,0 @@
-import { createOptimizedPicture } from '../../scripts/aem.js';
-
-/* export default function decorate(block) {
-  const ul = document.createElement('ul');
-  [...block.children].forEach((row) => {
-    const li = document.createElement('li');
-    while (row.firstElementChild) li.append(row.firstElementChild);
-    [...li.children].forEach((div) => {
-      const link = div.querySelectorAll('a[href]');
-      if (div.children.length === 1 && div.querySelector('picture')) div.className = 'cards-card-image';
-      else if (div.children.length === 1 && div.querySelector('p')) div.className = 'cards-card-title';
-      else div.className = 'cards-card-body';
-      if (link.length > 1) {
-        //
-      } else if (link.length === 1) {
-        const a = document.createElement('a');
-        a.href = link[0].href;
-        link[0].parentNode.remove();
-        a.append(div);
-        li.append(a);
-      }
-    });
-
-    ul.append(li);
-  });
-  ul.querySelectorAll('picture > img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
-  block.textContent = '';
-  block.append(ul);
-} */
-
-export default function decorate(block) {
-  /* change to ul, li */
-  const ul = document.createElement('ul');
-  [...block.children].forEach((row) => {
-    const li = document.createElement('li');
-    while (row.firstElementChild) li.append(row.firstElementChild);
-    [...li.children].forEach((div) => {
-      if (div.children.length === 1 && div.querySelector('picture')) div.className = 'cards-card-image';
-      else div.className = 'cards-card-body';
-    });
-    ul.append(li);
-  });
-  ul.querySelectorAll('picture > img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
-  block.textContent = '';
-  block.append(ul);
-}
