+++ fstab.yaml
@@ -1,2 +1,2 @@
 mountpoints:
-  /: https://adobe.sharepoint.com/sites/Rusher/Shared%20Documents/sites/bsc
+  /: https://adobe.sharepoint.com/sites/HelixProjects/Shared%20Documents/sites/blueshieldca
