+++ group/blocks/tabs/tabs.js
