+++ head.html
@@ -1,4 +1,4 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/scripts/aem.js" type="module"></script>
-<script src="/scripts/scripts.js" type="module"></script>
-<link rel="stylesheet" href="/styles/styles.css"/>
+<script src="/group/scripts/aem.js" type="module"></script>
+<script src="/group/scripts/scripts.js" type="module"></script>
+<link rel="stylesheet" href="/group/styles/styles.css"/>
