+++ package.json
@@ -4,8 +4,8 @@
   "version": "1.3.0",
   "description": "Starter project for Adobe Helix",
   "scripts": {
-    "lint:js": "eslint .",
-    "lint:css": "stylelint blocks/**/*.css styles/*.css",
+    "lint:js": "eslint . --ext .json,.js,.mjs",
+    "lint:css": "stylelint group/blocks/**/*.css group/styles/*.css",
     "lint": "npm run lint:js && npm run lint:css"
   },
   "repository": {
